fix exporting of hcgov AccordionItem so storybook & doc site have access to the prop override
